requirements.txt
colorama==0.4.6
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
openpyxl==3.1.2
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
scipy==1.7.3
six==1.16.0
tqdm==4.65.0
typing_extensions==4.7.0
